requirements.txt
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
billiard==4.2.0
celery==5.4.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
environs==8.0.0
et-xmlfile==1.1.0
executing==2.0.1
frozenlist==1.4.1
icecream==2.1.3
idna==3.4
kombu==5.3.7
magic-filter==1.0.12
marshmallow==3.21.1
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==11.3.0
pillow_heif==1.0.0
prompt_toolkit==3.0.46
psycopg2==2.9.9
pycparser==2.22
Pygments==2.17.2
pyheif==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
pytz==2023.3.post1
qrcode==8.1
redis==5.0.5
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
